requirements.txt
streamlit
transformers
torch
sentence-transformers
faiss-cpu
wikipedia
requests
beautifulsoup4
pandas
scikit-learn
numpy
python-dotenv
